working on post route

+++ db/schema/03_qandas.sql
@@ -8,7 +8,7 @@ CREATE TABLE question (
 DROP TABLE IF EXISTS answer CASCADE;
 CREATE TABLE answer (
   id SERIAL PRIMARY KEY NOT NULL,
-  question_id BIGINT REFERENCES question(id),
+  question_id INTEGER REFERENCES question(id),
   content TEXT,
   correct BOOLEAN
 );
